refractor: side scroller for course evaluations
